Refactor user functionalities to the controller
